refactor: ident, path
