+++ src/ast.rs
@@ -1,19 +1,6 @@
 use anyhow::{bail, Result};
 
-use crate::util::source::Source;
-
-#[derive(PartialEq, Debug, Clone, Hash, Eq)]
-pub struct Ident(pub String);
-
-impl Ident {
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-
-    pub fn to_string(&self) -> String {
-        self.0.clone()
-    }
-}
+use crate::util::{ident::Ident, source::Source};
 
 #[derive(PartialEq, Debug, Clone)]
 pub enum Type {
