+++ src/ir_code_gen.rs
@@ -3,10 +3,10 @@ use std::collections::HashMap;
 use anyhow::{anyhow, bail, Context, Result};
 
 use crate::{
-    ast::{Decl, Expr, Func, Ident, Lit, Module, Statement, Type},
+    ast::{Decl, Expr, Func, Lit, Module, Statement, Type},
     compiler::ErrorInSource,
     ir::{IrTerm, IrType},
-    util::source::Source,
+    util::{ident::Ident, source::Source},
 };
 
 #[derive(Debug, Clone)]
